Move out catalog query to core project
