+++ src/VirtoCommerce.ExperienceApiModule.Core/Index/ISearchDocumentsQuery.cs
@@ -1,6 +1,6 @@
 namespace VirtoCommerce.ExperienceApiModule.Core.Index
 {
-    public interface ISearchDocumentsQuery : ISearchQuery, IHaveObjectIds, IHaveSearchParams, IHasIncludeFields
+    public interface ISearchDocumentsQuery : ISearchQuery, IHaveObjectIds, IHaveSearchParams, ICatalogQuery
     {
     }
 }
